requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
kafka-python==2.0.2
requests==2.31.0
beautifulsoup4==4.12.2
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
psutil==5.9.6
pyspark==3.5.0